requirements.txt
yfinance>=0.2.32
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.2
